+++ bin/cmd.js
@@ -4,7 +4,8 @@ let exportBook = require("../export");
 const commandLineArgs = require("command-line-args");
 const commandLineUsage = require("command-line-usage")
 
-const optionDefinitions = [
+
+const parametersdefinitions =[
   {
     name: "book",
     alias: "b",
@@ -27,22 +28,36 @@ const optionDefinitions = [
     description: "Target output directory"
   },
   {
+    name: "help",
+    alias: "h",
+    type: Boolean,
+    description: "Displays this usage guide."
+  }
+];
+
+const optionDefinitions = [
+  {
     name: "binaries",
+    alias: "B",
     type: Boolean,
     description: "Export also referred binaries."
   },
   {
     name: "draws",
+    alias: "D",
     type: Boolean,
     description: "Export also referred UuBml.Draws."
   },
-
   {
-    name: "help",
-    alias: "h",
+    name: "transform",
+    alias: "T",
     type: Boolean,
-    description: "Displays this usage guide."
-  },
+    description: "Transform page body to formatted uu5string and stores it to another file."
+  }
+];
+const cliDefinition = [
+  ...parametersdefinitions,
+  ...optionDefinitions
 ];
 const sections = [
   {
@@ -52,9 +67,13 @@ const sections = [
   {
     header: "Synopsis",
     content: [
-      "uubookkit-exporter {bold --book} {underline url} {bold --token} {underline token} {bold --output} {underline dir} [{bold --binaries}] [{bold --draws}]",
+      "uubookkit-exporter [options] {bold --book} {underline url} {bold --token} {underline token} {bold --output} {underline dir}",
       "uubookkit-exporter {bold --help}"
     ]
+  },
+  {
+    header: "Parameters",
+    optionList: parametersdefinitions
   },
   {
     header: "Options",
@@ -62,7 +81,7 @@ const sections = [
   }
 ];
 const usage = commandLineUsage(sections);
-const options = commandLineArgs(optionDefinitions);
+const options = commandLineArgs(cliDefinition);
 
 const valid = options.help || (options.book && options.token && options.output);
 if (!valid || options.help) {
@@ -72,6 +91,7 @@ if (!valid || options.help) {
 
 let exportOptions = {
   exportBinaries: Boolean(options.binaries),
-  exportDraws: Boolean(options.draws)
+  exportDraws: Boolean(options.draws),
+  transformBody: Boolean(options.transform)
 }
 exportBook(options.book, options.token, options.output, exportOptions);