+++ export.js
@@ -26,8 +26,12 @@ const defaultOptions = {
   /**
    * Exports all Binaries used in book content.
    */
-  exportBinaries: false
-}
+  exportBinaries: false,
+  /**
+   * Transforms page body to formatted uu5string and stores it to another file.
+   */
+  transformBody: false
+};
 
 /**
  * Exports book source to specified directory.
@@ -44,6 +48,13 @@ async function exportBook(book, token, outputDir, options) {
   };
   logger.info(`Going to export book ${book}`);
   logger.info(`Using options : ${JSON.stringify(options)}`);
+
+  let exportDescriptor = {
+    book,
+    options: options,
+    exportStart: new Date(),
+    stats: {}
+  };
 
   let bookuri = UriBuilder.parse(book).clearParameters().setUseCase("").toUri();
   await mkdirp(outputDir);
@@ -64,18 +75,40 @@ async function exportBook(book, token, outputDir, options) {
   let otherResources = {
     binaries: new Set([]),
     draws: new Set([])
-  }
+  };
+  let pagesStats = {
+    exported: 0,
+    error: 0
+  };
   for (const page of pages) {
     const pageDataResponse = await AppClient.get(bookuri.toString() + "/loadPage", {code: page.code}, httpOptions);
     const pageData = pageDataResponse.data;
     _analyzePage(pageData, otherResources, options);
     const fileName = `${page.code}.json`;
     await writefile(path.join(pagesDir, fileName), JSON.stringify(pageData, null, 2), "utf8");
+    if (options.transformBody) {
+      let body = pageData.body;
+      if (!Array.isArray(body)) {
+        body = [body];
+      }
+      let transformedBody = body
+      .map(part => part.startsWith("<uu5string/>") ? part.substring("<uu5string/>".length) : part)
+      .join("\n\n<div hidden>Part end(uu5string does not support comments)</div>\n\n");
+      transformedBody = "<uu5string/>\n" + transformedBody;
+      await writefile(path.join(pagesDir, `${page.code}.uu5`), transformedBody, "utf8");
+      pagesStats.exported++;
+    }
   }
+  exportDescriptor.stats.pages = pagesStats;
+  exportDescriptor.stats.draws = await _downloadUuBmlDraws(outputDir, otherResources.draws, httpOptions);
+  exportDescriptor.stats.binaries = await _downloadBinaries(outputDir, otherResources.binaries, bookuri, httpOptions);
 
-  await _downloadUuBmlDraws(outputDir, otherResources.draws, httpOptions);
-  await _downloadBinaries(outputDir, otherResources.binaries, bookuri, httpOptions);
+  exportDescriptor.exportEnd = new Date();
+  exportDescriptor.stats.duration = exportDescriptor.exportEnd - exportDescriptor.exportStart;
 
+  await writefile(path.join(outputDir, "export-result.json"), JSON.stringify(exportDescriptor, null, 2), "utf8");
+  logger.info("Export has been finished. See export statistics bellow.");
+  console.log(JSON.stringify(exportDescriptor, null, 2));
 }
 
 /**
@@ -87,16 +120,23 @@ async function exportBook(book, token, outputDir, options) {
  * @private
  */
 _downloadUuBmlDraws = async function _downloadUuBmlDraws(outputDir, draws, httpOptions) {
+  let stats = {
+    exported: 0,
+    error: 0
+  };
   let drawsDir = path.join(outputDir, "draws");
   await mkdirp(drawsDir);
   for (let draw of draws) {
     try {
       logger.info(`Downloading UuBmlDraw ${draw}.`);
       await downloadUuBmlDraw(draw, httpOptions, drawsDir);
+      stats.exported++;
     } catch (e) {
+      stats.error++;
       logger.error(`UuBmlDraw ${draw} cannot be downloaded.`, e);
     }
   }
+  return stats;
 }
 
 /**
@@ -109,6 +149,11 @@ _downloadUuBmlDraws = async function _downloadUuBmlDraws(outputDir, draws, httpO
  * @private
  */
 _downloadBinaries = async function _downloadBinaries(outputDir, binaries, bookuri, httpOptions) {
+  let stats = {
+    exported: 0,
+    notFound: 0,
+    error: 0
+  };
   let binariesDir = path.join(outputDir, "binaries");
   await mkdirp(binariesDir);
   for (let binary of binaries) {
@@ -118,14 +163,18 @@ _downloadBinaries = async function _downloadBinaries(outputDir, binaries, bookur
       let response = await AppClient.get(uri, null, httpOptions);
       let binaryPath = path.join(binariesDir, binary);
       await response.data.pipe(fs.createWriteStream(binaryPath));
+      stats.exported++;
     } catch (e) {
       if (e.code === "uu-app-binarystore/uuBinaryGetBinaryData/uuBinaryDoesNotExist") {
+        stats.notFound++;
         logger.warn(`Binary ${binary} does not exist.`);
       } else {
+        stats.error++;
         logger.error(`Binary ${binary} cannot be downloaded.`, e);
       }
     }
   }
+  return stats;
 }
 
 function _processNode(node, ctx, options) {