+++ doc/CHANGELOG.md
@@ -1,3 +1,8 @@
+0.3.1
+-----
+* Added possibility to transform page content to formatted uu5string in another file.
+* Added export-result.json with basic information about export.
+
 0.3.0
 -----
 * Added possibility to export referred Binaries and UuBml.Draws.